refactor: extract content_hint module for generic output saving
- create src/core/content_hint.rs with generic save_output* functions
- add RTK_HINTS env var support (RTK_TEE still works for backward compat)
- update json_cmd.rs to use save_output_and_hint for truncation recovery
- update tee.rs to delegate to content_hint, remove dead code
- migrate duplicate tests to content_hint.rs
- remove MIN_TEE_SIZE alias, use MIN_CONTENT_SIZE directly
