+++ src/core/mod.rs
@@ -2,6 +2,7 @@
 
 pub mod config;
 pub mod constants;
+pub mod content_hint;
 pub mod display_helpers;
 pub mod filter;
 pub mod runner;
