+++ src/core/tee.rs
@@ -1,224 +1,19 @@
 //! Raw output recovery -- saves unfiltered output to disk on command failure.
+//! This module is kept for backward compatibility.
+//! New code should use content_hint module instead.
 
-use super::constants::RTK_DATA_DIR;
-use crate::core::config::Config;
+use crate::core::content_hint as ch;
 use std::path::PathBuf;
-
-/// Minimum output size to tee (smaller outputs don't need recovery)
-const MIN_TEE_SIZE: usize = 500;
-
-/// Default max files to keep in tee directory
-const DEFAULT_MAX_FILES: usize = 20;
-
-/// Default max file size (1MB)
-const DEFAULT_MAX_FILE_SIZE: usize = 1_048_576;
-
-/// Sanitize a command slug for use in filenames.
-/// Replaces non-alphanumeric chars (except underscore/hyphen) with underscore,
-/// truncates at 40 chars.
-fn sanitize_slug(slug: &str) -> String {
-    let sanitized: String = slug
-        .chars()
-        .map(|c| {
-            if c.is_ascii_alphanumeric() || c == '_' || c == '-' {
-                c
-            } else {
-                '_'
-            }
-        })
-        .collect();
-    if sanitized.len() > 40 {
-        sanitized[..40].to_string()
-    } else {
-        sanitized
-    }
-}
-
-/// Get the tee directory, respecting config and env overrides.
-fn get_tee_dir(config: &Config) -> Option<PathBuf> {
-    // Env var override
-    if let Ok(dir) = std::env::var("RTK_TEE_DIR") {
-        return Some(PathBuf::from(dir));
-    }
-
-    // Config override
-    if let Some(ref dir) = config.tee.directory {
-        return Some(dir.clone());
-    }
-
-    // Default: ~/.local/share/rtk/tee/
-    dirs::data_local_dir().map(|d| d.join(RTK_DATA_DIR).join("tee"))
-}
-
-/// Rotate old tee files: keep only the last `max_files`, delete oldest.
-fn cleanup_old_files(dir: &std::path::Path, max_files: usize) {
-    let mut entries: Vec<_> = std::fs::read_dir(dir)
-        .ok()
-        .into_iter()
-        .flatten()
-        .filter_map(|e| e.ok())
-        .filter(|e| e.path().extension().is_some_and(|ext| ext == "log"))
-        .collect();
-
-    if entries.len() <= max_files {
-        return;
-    }
-
-    // Sort by filename (which starts with epoch timestamp = chronological)
-    entries.sort_by_key(|e| e.file_name());
-
-    let to_remove = entries.len() - max_files;
-    for entry in entries.iter().take(to_remove) {
-        let _ = std::fs::remove_file(entry.path());
-    }
-}
-
-/// Check if tee should be skipped based on config, mode, exit code, and size.
-/// Returns None if should skip, Some(tee_dir) if should proceed.
-fn should_tee(
-    config: &TeeConfig,
-    raw_len: usize,
-    exit_code: i32,
-    tee_dir: Option<PathBuf>,
-) -> Option<PathBuf> {
-    if !config.enabled {
-        return None;
-    }
-
-    match config.mode {
-        TeeMode::Never => return None,
-        TeeMode::Failures => {
-            if exit_code == 0 {
-                return None;
-            }
-        }
-        TeeMode::Always => {}
-    }
-
-    if raw_len < MIN_TEE_SIZE {
-        return None;
-    }
-
-    tee_dir
-}
-
-/// Write raw output to a tee file in the given directory.
-/// Returns file path on success.
-fn write_tee_file(
-    raw: &str,
-    command_slug: &str,
-    tee_dir: &std::path::Path,
-    max_file_size: usize,
-    max_files: usize,
-) -> Option<PathBuf> {
-    std::fs::create_dir_all(tee_dir).ok()?;
-
-    let slug = sanitize_slug(command_slug);
-    let epoch = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .ok()?
-        .as_secs();
-    let filename = format!("{}_{}.log", epoch, slug);
-    let filepath = tee_dir.join(filename);
-
-    // Truncate at max_file_size (find a safe UTF-8 char boundary)
-    let content = if raw.len() > max_file_size {
-        let boundary = raw
-            .char_indices()
-            .take_while(|(i, _)| *i < max_file_size)
-            .last()
-            .map(|(i, c)| i + c.len_utf8())
-            .unwrap_or(0);
-        format!(
-            "{}\n\n--- truncated at {} bytes ---",
-            &raw[..boundary],
-            max_file_size
-        )
-    } else {
-        raw.to_string()
-    };
-
-    std::fs::write(&filepath, content).ok()?;
-
-    // Rotate old files
-    cleanup_old_files(tee_dir, max_files);
-
-    Some(filepath)
-}
-
-/// Write raw output to tee file if conditions are met.
-/// Returns file path on success, None if skipped/failed.
-pub fn tee_raw(raw: &str, command_slug: &str, exit_code: i32) -> Option<PathBuf> {
-    // Check RTK_TEE=0 env override (disable)
-    if std::env::var("RTK_TEE").ok().as_deref() == Some("0") {
-        return None;
-    }
-
-    let config = Config::load().ok()?;
-    let tee_dir = get_tee_dir(&config)?;
-
-    let tee_dir = should_tee(&config.tee, raw.len(), exit_code, Some(tee_dir))?;
-
-    write_tee_file(
-        raw,
-        command_slug,
-        &tee_dir,
-        config.tee.max_file_size,
-        config.tee.max_files,
-    )
-}
-
-fn display_path(path: &std::path::Path) -> String {
-    if let Some(home) = dirs::home_dir() {
-        if let Ok(relative) = path.strip_prefix(&home) {
-            return format!("~/{}", relative.display());
-        }
-    }
-    path.display().to_string()
-}
-
-fn format_hint(path: &std::path::Path) -> String {
-    format!("[full output: {}]", display_path(path))
-}
 
 /// Convenience: tee + format hint in one call.
 /// Returns hint string if file was written, None if skipped.
 pub fn tee_and_hint(raw: &str, command_slug: &str, exit_code: i32) -> Option<String> {
-    let path = tee_raw(raw, command_slug, exit_code)?;
-    Some(format_hint(&path))
-}
-
-fn force_tee_path(content: &str, command_slug: &str) -> Option<PathBuf> {
-    if std::env::var("RTK_TEE").ok().as_deref() == Some("0") {
-        return None;
-    }
-
-    if content.is_empty() {
-        return None;
-    }
-
-    let config = Config::load().ok()?;
-
-    if !config.tee.enabled {
-        return None;
-    }
-
-    let tee_dir = get_tee_dir(&config)?;
-    let tee_dir = std::fs::create_dir_all(&tee_dir).ok().and(Some(tee_dir))?;
-
-    write_tee_file(
-        content,
-        command_slug,
-        &tee_dir,
-        config.tee.max_file_size,
-        config.tee.max_files,
-    )
+    ch::save_output_on_failure(raw, command_slug, ".log", exit_code)
 }
 
 /// Returns `[full output: ~/path]`, or None if tee is disabled/skipped.
 pub fn force_tee_hint(raw: &str, command_slug: &str) -> Option<String> {
-    let path = force_tee_path(raw, command_slug)?;
-    Some(format_hint(&path))
+    ch::save_output_and_hint(raw, command_slug, ".log")
 }
 
 /// Returns `[see remaining: tail -n +{line_offset} ~/path]`, or None if tee is disabled/skipped.
@@ -227,11 +22,11 @@ pub fn force_tee_tail_hint(
     command_slug: &str,
     line_offset: usize,
 ) -> Option<String> {
-    let path = force_tee_path(content, command_slug)?;
+    let path = ch::save_output(content, command_slug, ".log")?;
     Some(format!(
         "[see remaining: tail -n +{} {}]",
         line_offset,
-        display_path(&path)
+        ch::display_path(&path)
     ))
 }
 
@@ -261,8 +56,8 @@ impl Default for TeeConfig {
         Self {
             enabled: true,
             mode: TeeMode::default(),
-            max_files: DEFAULT_MAX_FILES,
-            max_file_size: DEFAULT_MAX_FILE_SIZE,
+            max_files: ch::DEFAULT_MAX_FILES,
+            max_file_size: ch::DEFAULT_MAX_FILE_SIZE,
             directory: None,
         }
     }
@@ -271,181 +66,6 @@ impl Default for TeeConfig {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use std::fs;
-
-    #[test]
-    fn test_sanitize_slug() {
-        assert_eq!(sanitize_slug("cargo_test"), "cargo_test");
-        assert_eq!(sanitize_slug("cargo test"), "cargo_test");
-        assert_eq!(sanitize_slug("cargo-test"), "cargo-test");
-        assert_eq!(sanitize_slug("go/test/./pkg"), "go_test___pkg");
-        // Truncate at 40
-        let long = "a".repeat(50);
-        assert_eq!(sanitize_slug(&long).len(), 40);
-    }
-
-    #[test]
-    fn test_should_tee_disabled() {
-        let config = TeeConfig {
-            enabled: false,
-            ..TeeConfig::default()
-        };
-        let dir = PathBuf::from("/tmp/tee");
-        assert!(should_tee(&config, 1000, 1, Some(dir)).is_none());
-    }
-
-    #[test]
-    fn test_should_tee_never_mode() {
-        let config = TeeConfig {
-            mode: TeeMode::Never,
-            ..TeeConfig::default()
-        };
-        let dir = PathBuf::from("/tmp/tee");
-        assert!(should_tee(&config, 1000, 1, Some(dir)).is_none());
-    }
-
-    #[test]
-    fn test_should_tee_skip_small_output() {
-        let config = TeeConfig::default();
-        let dir = PathBuf::from("/tmp/tee");
-        // Below MIN_TEE_SIZE (500)
-        assert!(should_tee(&config, 100, 1, Some(dir)).is_none());
-    }
-
-    #[test]
-    fn test_should_tee_skip_success_in_failures_mode() {
-        let config = TeeConfig::default(); // mode = Failures
-        let dir = PathBuf::from("/tmp/tee");
-        assert!(should_tee(&config, 1000, 0, Some(dir)).is_none());
-    }
-
-    #[test]
-    fn test_should_tee_proceed_on_failure() {
-        let config = TeeConfig::default(); // mode = Failures
-        let dir = PathBuf::from("/tmp/tee");
-        assert!(should_tee(&config, 1000, 1, Some(dir)).is_some());
-    }
-
-    #[test]
-    fn test_should_tee_always_mode_success() {
-        let config = TeeConfig {
-            mode: TeeMode::Always,
-            ..TeeConfig::default()
-        };
-        let dir = PathBuf::from("/tmp/tee");
-        assert!(should_tee(&config, 1000, 0, Some(dir)).is_some());
-    }
-
-    #[test]
-    fn test_write_tee_file_creates_file() {
-        let tmpdir = tempfile::tempdir().unwrap();
-        let content = "error: test failed\n".repeat(50);
-        let result = write_tee_file(
-            &content,
-            "cargo_test",
-            tmpdir.path(),
-            DEFAULT_MAX_FILE_SIZE,
-            20,
-        );
-        assert!(result.is_some());
-
-        let path = result.unwrap();
-        assert!(path.exists());
-        let written = fs::read_to_string(&path).unwrap();
-        assert!(written.contains("error: test failed"));
-    }
-
-    #[test]
-    fn test_write_tee_file_truncation() {
-        let tmpdir = tempfile::tempdir().unwrap();
-        let big_output = "x".repeat(2000);
-        // Set max_file_size to 1000 bytes
-        let result = write_tee_file(&big_output, "test", tmpdir.path(), 1000, 20);
-        assert!(result.is_some());
-
-        let path = result.unwrap();
-        let content = fs::read_to_string(&path).unwrap();
-        assert!(content.contains("--- truncated at 1000 bytes ---"));
-        assert!(content.len() < 2000);
-    }
-
-    #[test]
-    fn test_write_tee_file_truncation_utf8_boundary() {
-        let tmpdir = tempfile::tempdir().unwrap();
-        // Create a string where the truncation point falls inside a multi-byte char.
-        // Japanese chars are 3 bytes each in UTF-8.
-        // 332 chars * 3 bytes = 996 bytes, then one more = 999 bytes.
-        // With max_file_size=998, the cut falls mid-character.
-        let japanese = "\u{6F22}".repeat(333); // 999 bytes of 3-byte chars
-        assert_eq!(japanese.len(), 999);
-
-        // Truncate at 998 — falls in the middle of the 333rd character
-        let result = write_tee_file(&japanese, "test_utf8", tmpdir.path(), 998, 20);
-        assert!(result.is_some());
-
-        let path = result.unwrap();
-        let content = fs::read_to_string(&path).unwrap();
-        assert!(content.contains("--- truncated at 998 bytes ---"));
-        // Should contain 332 full characters (996 bytes), not panic
-        assert!(content.starts_with(&"\u{6F22}".repeat(332)));
-    }
-
-    #[test]
-    fn test_write_tee_file_truncation_emoji() {
-        let tmpdir = tempfile::tempdir().unwrap();
-        // Emoji are 4 bytes each in UTF-8
-        let emojis = "\u{1F600}".repeat(100); // 400 bytes
-        assert_eq!(emojis.len(), 400);
-
-        // Truncate at 201 — falls mid-emoji (4-byte boundary is at 200, 204)
-        let result = write_tee_file(&emojis, "test_emoji", tmpdir.path(), 201, 20);
-        assert!(result.is_some());
-
-        let path = result.unwrap();
-        let content = fs::read_to_string(&path).unwrap();
-        assert!(content.contains("--- truncated at 201 bytes ---"));
-        // The emoji portion should be exactly 200 bytes (50 emojis),
-        // rounded down from 201 to the nearest char boundary
-        let target = "\u{1F600}".repeat(50);
-        assert!(content.starts_with(&target));
-    }
-
-    #[test]
-    fn test_cleanup_old_files() {
-        let tmpdir = tempfile::tempdir().unwrap();
-        let dir = tmpdir.path();
-
-        // Create 25 .log files
-        for i in 0..25 {
-            let filename = format!("{:010}_{}.log", 1000000 + i, "test");
-            fs::write(dir.join(&filename), "content").unwrap();
-        }
-
-        cleanup_old_files(dir, 20);
-
-        let remaining: Vec<_> = fs::read_dir(dir).unwrap().filter_map(|e| e.ok()).collect();
-        assert_eq!(remaining.len(), 20);
-
-        // Oldest 5 should be removed
-        for i in 0..5 {
-            let filename = format!("{:010}_{}.log", 1000000 + i, "test");
-            assert!(!dir.join(&filename).exists());
-        }
-        // Newest 20 should remain
-        for i in 5..25 {
-            let filename = format!("{:010}_{}.log", 1000000 + i, "test");
-            assert!(dir.join(&filename).exists());
-        }
-    }
-
-    #[test]
-    fn test_format_hint() {
-        let path = PathBuf::from("/tmp/rtk/tee/123_cargo_test.log");
-        let hint = format_hint(&path);
-        assert!(hint.starts_with("[full output: "));
-        assert!(hint.ends_with(']'));
-        assert!(hint.contains("123_cargo_test.log"));
-    }
 
     #[test]
     fn test_tee_config_default() {
@@ -518,7 +138,7 @@ directory = "/tmp/rtk-tee"
     #[test]
     fn test_force_tee_tail_hint_format() {
         let path = std::path::PathBuf::from("/tmp/rtk/tee/123_docker_images.log");
-        let display = display_path(&path);
+        let display = ch::display_path(&path);
         let hint = format!("[see remaining: tail -n +{} {}]", 22, display);
         assert!(hint.starts_with("[see remaining: tail -n +22 "));
         assert!(hint.ends_with(']'));
